auth.sh: fix per user race
So far we had 1 cache file per user, but users can log in parallel with
the same user.

Fix by using a proper temporary file. As this already uses the final
director, the mv is atomic. While at it use a "trap" and a "rm -f" (the
file does not need to exist in the first place).